requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==1.4.32
typing_extensions==4.9.0
Werkzeug==3.0.1
